# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "EDCOCR API",
    "description": "REST API for forensic-grade document OCR processing.",
    "version": "4.1.0"
  },
  "paths": {
    "/api/v1/jobs/batch": {
      "post": {
        "tags": [
          "batch"
        ],
        "summary": "Submit Batch",
        "description": "Submit multiple documents for OCR processing as a batch.",
        "operationId": "submit_batch_api_v1_jobs_batch_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_submit_batch_api_v1_jobs_batch_post"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchSubmitResponse"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/api__models__ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/api__models__ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/api__models__ErrorResponse"
                }
              }
            },
            "description": "Too Many Requests"
          }
        }
      },
      "get": {
        "tags": [
          "batch"
        ],
        "summary": "List Batches",
        "description": "List batches with optional status filter and pagination.",
        "operationId": "list_batches_api_v1_jobs_batch_get",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "per_page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Per Page"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/jobs/batch/{batch_id}": {
      "get": {
        "tags": [
          "batch"
        ],
        "summary": "Get Batch Status",
        "description": "Get status and progress of a batch.",
        "operationId": "get_batch_status_api_v1_jobs_batch__batch_id__get",
        "parameters": [
          {
            "name": "batch_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Batch Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchStatusResponse"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/api__models__ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "batch"
        ],
        "summary": "Cancel Batch",
        "description": "Cancel all running jobs in a batch.",
        "operationId": "cancel_batch_api_v1_jobs_batch__batch_id__delete",
        "parameters": [
          {
            "name": "batch_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Batch Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchStatusResponse"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/api__models__ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/jobs/batch/{batch_id}/retry": {
      "post": {
        "tags": [
          "batch"
        ],
        "summary": "Retry Batch",
        "description": "Retry all failed or cancelled jobs in a batch.",
        "operationId": "retry_batch_api_v1_jobs_batch__batch_id__retry_post",
        "parameters": [
          {
            "name": "batch_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Batch Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchStatusResponse"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/api__models__ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/api__models__ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BatchJobSummary": {
        "properties": {
          "job_id": {
            "type": "string",
            "title": "Job Id"
          },
          "source_file": {
            "type": "string",
            "title": "Source File"
          },
          "status": {
            "type": "string",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "job_id",
          "source_file",
          "status"
        ],
        "title": "BatchJobSummary",
        "description": "Summary of a single job within a batch."
      },
      "BatchLinks": {
        "properties": {
          "self": {
            "type": "string",
            "title": "Self"
          },
          "jobs": {
            "type": "string",
            "title": "Jobs"
          }
        },
        "type": "object",
        "required": [
          "self",
          "jobs"
        ],
        "title": "BatchLinks",
        "description": "HATEOAS links for batch responses."
      },
      "BatchListResponse": {
        "properties": {
          "batches": {
            "items": {
              "$ref": "#/components/schemas/BatchStatusResponse"
            },
            "type": "array",
            "title": "Batches"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "per_page": {
            "type": "integer",
            "title": "Per Page"
          }
        },
        "type": "object",
        "required": [
          "batches",
          "total",
          "page",
          "per_page"
        ],
        "title": "BatchListResponse",
        "description": "Paginated list of batches."
      },
      "BatchProgressInfo": {
        "properties": {
          "submitted": {
            "type": "integer",
            "title": "Submitted",
            "default": 0
          },
          "processing": {
            "type": "integer",
            "title": "Processing",
            "default": 0
          },
          "completed": {
            "type": "integer",
            "title": "Completed",
            "default": 0
          },
          "failed": {
            "type": "integer",
            "title": "Failed",
            "default": 0
          },
          "cancelled": {
            "type": "integer",
            "title": "Cancelled",
            "default": 0
          },
          "percent_complete": {
            "type": "number",
            "title": "Percent Complete",
            "default": 0.0
          }
        },
        "type": "object",
        "title": "BatchProgressInfo",
        "description": "Aggregated progress information for a batch."
      },
      "BatchStatusResponse": {
        "properties": {
          "batch_id": {
            "type": "string",
            "title": "Batch Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "processing_time": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Processing Time"
          },
          "total_jobs": {
            "type": "integer",
            "title": "Total Jobs"
          },
          "progress": {
            "$ref": "#/components/schemas/BatchProgressInfo"
          },
          "jobs": {
            "items": {
              "$ref": "#/components/schemas/BatchJobSummary"
            },
            "type": "array",
            "title": "Jobs"
          },
          "settings": {
            "additionalProperties": true,
            "type": "object",
            "title": "Settings",
            "default": {}
          },
          "webhook_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Webhook Status"
          }
        },
        "type": "object",
        "required": [
          "batch_id",
          "status",
          "created_at",
          "total_jobs",
          "progress",
          "jobs"
        ],
        "title": "BatchStatusResponse",
        "description": "Full batch status with progress and child job summaries."
      },
      "BatchSubmitResponse": {
        "properties": {
          "batch_id": {
            "type": "string",
            "title": "Batch Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "total_jobs": {
            "type": "integer",
            "title": "Total Jobs"
          },
          "priority": {
            "type": "string",
            "title": "Priority"
          },
          "jobs": {
            "items": {
              "$ref": "#/components/schemas/BatchJobSummary"
            },
            "type": "array",
            "title": "Jobs"
          },
          "links": {
            "$ref": "#/components/schemas/BatchLinks"
          }
        },
        "type": "object",
        "required": [
          "batch_id",
          "status",
          "created_at",
          "total_jobs",
          "priority",
          "jobs",
          "links"
        ],
        "title": "BatchSubmitResponse",
        "description": "Response returned when a batch is successfully submitted."
      },
      "Body_submit_batch_api_v1_jobs_batch_post": {
        "properties": {
          "files": {
            "items": {
              "type": "string",
              "format": "binary"
            },
            "type": "array",
            "title": "Files",
            "description": "List of files to upload and process.",
            "default": []
          },
          "source_paths": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Paths",
            "description": "JSON array string of server-side file paths."
          },
          "priority": {
            "type": "string",
            "title": "Priority",
            "description": "Processing priority: 'low', 'normal', or 'urgent'.",
            "default": "normal"
          },
          "enable_docintel": {
            "type": "boolean",
            "title": "Enable Docintel",
            "description": "Enable structured Document Intelligence extraction (requires GPU).",
            "default": false
          },
          "docintel_mode": {
            "type": "string",
            "title": "Docintel Mode",
            "description": "DocIntel extraction mode: 'layout_only', 'tables_only', or 'full'.",
            "default": "full"
          },
          "skip_ocr": {
            "type": "boolean",
            "title": "Skip Ocr",
            "description": "Skip OCR and perform NLP/DocIntel only. Assumes textual input.",
            "default": false
          },
          "processing_timeout_minutes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Processing Timeout Minutes",
            "description": "Custom timeout in minutes for these jobs before they are marked failed."
          },
          "webhook_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Webhook Url",
            "description": "URL to receive a POST request when the batch completes or fails."
          },
          "webhook_secret": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Webhook Secret",
            "description": "Secret used to sign the webhook payload (HMAC-SHA256)."
          }
        },
        "type": "object",
        "title": "Body_submit_batch_api_v1_jobs_batch_post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "api__models__ErrorResponse": {
        "properties": {
          "error": {
            "type": "string",
            "title": "Error"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "details": {
            "additionalProperties": true,
            "type": "object",
            "title": "Details",
            "default": {}
          }
        },
        "type": "object",
        "required": [
          "error",
          "message"
        ],
        "title": "ErrorResponse"
      }
    },
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "API key for authentication. Set via OCR_API_KEY environment variable."
      },
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "OAuth2 Bearer token (when OAUTH2_ENABLED=true)."
      }
    }
  },
...
  "security": [
    {
      "ApiKeyAuth": []
    },
    {
      "BearerAuth": []
    }
  ]
}